Update JS filed and CSS
