+++ src/css/base.css
@@ -51,6 +51,5 @@ body {
     height: 100%;
     overflow: scroll;
     position: relative;
-    padding-top: 20vh;
   }
 }
