+++ src/index.html
@@ -11,12 +11,8 @@
       media="screen"
       href="./css/main.css"
     />
-    <link
-      rel="icon"
-      href="./images/bxRadio_mark.ico"
-      type="image/x-icon"
-    />
-    <script src="./js/app.js" defer></script>
+    <link rel="icon" href="./images/bxRadio_mark.ico" type="image/x-icon" />
+    <script src="./js/app.js" defer type="module"></script>
   </head>
 
   <body>
